# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: LOC-DB Central Component
host: localhost
basePath: /
...
schemes:
  #- http
  - https
consumes:
  - application/json
  - multipart/form-data
produces:
  - application/json
  - image/png
  - application/pdf
securityDefinitions:
  basicAuth:
    type: basic
  login:
    type: basic

# To apply Basic auth to the whole API:
#security:
#  - basicAuth: []
paths:
...
  /addTargetBibliographicResource:
    x-swagger-router-controller: bibliographicEntry
    get:
      tags: [BibliographicEntry]
      description: |
        Given the _id of a bibliographic entry and the _id of a target, i.e., a cited, bibliographic resource,
        this service takes care of setting all related properties in a consistent way. More precisely, the following three things happen:
        1) The 'status' of the bibliographic entry will be changed to VALID
        2) The 'references' property of the bibliographic entry will be set to the _id of the target bibliographic resource
        3) The 'cites' property of the parent resource of the entry will be extended with the _id of the target bibliographic reource
      operationId: addTargetBibliographicResource
      parameters:
        - name: bibliographicEntryId
          in: query
          description: The _id of the subject bibliographic entry that should point to the target bibliographic resource
          required: true
          type: string
        - name: bibliographicResourceId
          in: query
          description: The _id of the target bibliographic resource which the bibliographic entry should point to
          required: true
          type: string
      responses:
        "200":
          description: Success; In case the operation suceeds, the updated source bibliographic resource including the respective bibliographic entry will be returned
          schema:
            $ref: "#/definitions/BibliographicResource"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /removeTargetBibliographicResource:
    x-swagger-router-controller: bibliographicEntry
    get:
      tags: [BibliographicEntry]
      description: |
        Given the _id of a bibliographic entry this service clears all properties related to a potential target resource in a consistent way. More precisely, the following three things happen:
        1) The 'references' property of the bibliographic entry will be cleared
        2) The _id of the old target resource will be removed from teh 'cites' property of the parent resource of the entry
        3) The 'status' of the bibliographic entry will be changed to OCR_PROCESSED --> Is this always ok?
        Note that the old target bibliographic resource will not be deleted.
      operationId: removeTargetBibliographicResource
      parameters:
        - name: bibliographicEntryId
          in: query
          description: The _id of the subject bibliographic entry that should point to the target bibliographic resource
          required: true
          type: string
      responses:
        "200":
          description: Success; In case the operation suceeds, the updated source bibliographic resource including the respective bibliographic entry will be returned
          schema:
            $ref: "#/definitions/BibliographicResource"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /getToDoBibliographicEntries:
    x-swagger-router-controller: bibliographicEntry
    get:
      tags: [BibliographicEntry]
      description: Retrieves the BEs that are not processed by a librarian yet.
      operationId: getToDoBibliographicEntries
      parameters:
        - name: scanId
          in: query
          description: The Id of the Scan for which the BEs shall be retrieved
          required: false
          type: string
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/BibliographicEntry"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /getInternalSuggestionsByQueryString:
    x-swagger-router-controller: bibliographicEntry
    get:
      tags: [BibliographicEntry]
      description: Retrieves internal br suggestions for a given query string related to a bibliographic entry.
      operationId: getInternalSuggestionsByQueryString
      parameters:
        - name: query
          in: query
          description: The query string for which the suggestions shall be retrieved
          required: true
          type: string
        - name: threshold
          in: query
          description: The elastic relevance score for above which entrys shall be retrieved
          required: false
          type: number
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/BibliographicResource"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /getExternalSuggestionsByQueryString:
    x-swagger-router-controller: bibliographicEntry
    get:
      tags: [BibliographicEntry]
      description: Retrieves external br suggestions for a given query string related to a bibliographic entry.
      operationId: getExternalSuggestionsByQueryString
      parameters:
        - name: query
          in: query
          description: The query string for which the suggestions shall be retrieved
          required: true
          type: string
        - name: threshold
          in: query
          description: The sring similarity (dice coefficient) between the query string and the title + subtile + contributors above which suggestions shall be retrieved
          required: false
          type: number
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/BibliographicResource"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
...
  Identifier:
    required:
      - scheme
    properties:
      literalValue:
        type: string
      scheme:
        type: string
...
  BibliographicEntry:
    properties:
      _id:
        type: string
      identifiers:
        type: array
        items:
          $ref: "#/definitions/Identifier"
      bibliographicEntryText:
        type: string
      references:
        type: string
      scanId:
        type: string
      status:
        type: string
      ocrData:
        $ref: "#/definitions/OCRData"
  OCRData:
    properties:
      coordinates:
        type: string
      authors:
        type: array
        items:
          type: string
      title:
        type: string
      date:
        type: string
      marker:
        type: string
      comments:
        type: string
      journal:
        type: string
      volume:
        type: string
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...